// src/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type Rating = {
  __typename: "Rating",
  id: string,
  rating?: number | null,
  comment?: string | null,
  driver_image?: string | null,
  driver_name?: string | null,
  customer_name?: string | null,
  customer_image?: string | null,
  tripsID: string,
  driverID: string,
  userID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};
// ...
export type TripCart = {
  __typename: "TripCart",
  id: string,
  expiry_date: string,
  name: string,
  image: string,
  quantity: number,
  price: number,
  description?: string | null,
  storeID: string,
  userID: string,
  tripsID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};
// ...
export enum TripStatus {
  NEW = "NEW",
  ACCEPTED = "ACCEPTED",
  PICKED_UP = "PICKED_UP",
  COMPLETED = "COMPLETED",
  CANCELED = "CANCELED",
}
// ...
export type ModelTripStatusInput = {
  eq?: TripStatus | null,
  ne?: TripStatus | null,
};

export type Trips = {
  __typename: "Trips",
  id: string,
  trip_status?: TripStatus | null,
  pickup_lat?: string | null,
  pickup_lng?: string | null,
  dropoff_lat?: string | null,
  dropoff_lng?: string | null,
  pickup_location_description?: string | null,
  dropoff_location_description?: string | null,
  trip_cost?: number | null,
  TotalQtyItems?: number | null,
  receipt?: string | null,
  duration?: number | null,
  distance?: number | null,
  store_name?: string | null,
  store_image?: string | null,
  user_name?: string | null,
  user_photo?: string | null,
  mobile_number?: string | null,
  driver_name?: string | null,
  driver_photo?: string | null,
  driver_rating?: number | null,
  driver_mobile_number?: string | null,
  driver_car_model?: string | null,
  driver_car_make?: string | null,
  driver_car_color?: string | null,
  driver_car_plate_number?: string | null,
  sub_total?: number | null,
  redeem_points?: number | null,
  service_fee?: number | null,
  delivery_fee?: number | null,
  delivery_discount?: number | null,
  sum_total?: number | null,
  Driver?: Driver | null,
  userID?: string | null,
  storeID: string,
  TripCarts?: ModelTripCartConnection | null,
  Ratings?: ModelRatingConnection | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  tripsDriverId?: string | null,
};

export type Driver = {
  __typename: "Driver",
  id: string,
  name: string,
  email: string,
  image?: string | null,
  phone_number?: string | null,
  rating?: number | null,
  inviteCode?: string | null,
  lng?: string | null,
  lat?: string | null,
  verificationStatus?: DriverRegistrationStatus | null,
  onlineStatus?: boolean | null,
  Cars?: ModelCarConnection | null,
  Ratings?: ModelRatingConnection | null,
  DriverWallets?: ModelDriverWalletConnection | null,
  LicenseDocuments?: ModelLicenseDocumentConnection | null,
  RegistrationDocuments?: ModelRegistrationDocumentConnection | null,
  InsuranceDocuments?: ModelInsuranceDocumentConnection | null,
  fcmToken?: string | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export enum DriverRegistrationStatus {
  NEW = "NEW",
  COMPLETE = "COMPLETE",
  REJECTED = "REJECTED",
}


export type ModelCarConnection = {
  __typename: "ModelCarConnection",
  items:  Array<Car | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelRatingConnection = {
  __typename: "ModelRatingConnection",
  items:  Array<Rating | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelDriverWalletConnection = {
  __typename: "ModelDriverWalletConnection",
  items:  Array<DriverWallet | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelLicenseDocumentConnection = {
  __typename: "ModelLicenseDocumentConnection",
  items:  Array<LicenseDocument | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type LicenseDocument = {
  __typename: "LicenseDocument",
  id: string,
  image?: string | null,
  name?: string | null,
  expiry_date?: string | null,
  driverID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type ModelRegistrationDocumentConnection = {
  __typename: "ModelRegistrationDocumentConnection",
  items:  Array<RegistrationDocument | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type RegistrationDocument = {
  __typename: "RegistrationDocument",
  id: string,
  image?: string | null,
  name?: string | null,
  expiry_date?: string | null,
  driverID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type ModelInsuranceDocumentConnection = {
  __typename: "ModelInsuranceDocumentConnection",
  items:  Array<InsuranceDocument | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type InsuranceDocument = {
  __typename: "InsuranceDocument",
  id: string,
  image?: string | null,
  name?: string | null,
  expiry_date?: string | null,
  driverID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type ModelTripCartConnection = {
  __typename: "ModelTripCartConnection",
  items:  Array<TripCart | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};
// ...
export type ModelTripsConnection = {
  __typename: "ModelTripsConnection",
  items:  Array<Trips | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};
// ...
export type ModelRatingFilterInput = {
  id?: ModelIDInput | null,
  rating?: ModelIntInput | null,
  comment?: ModelStringInput | null,
  driver_image?: ModelStringInput | null,
  driver_name?: ModelStringInput | null,
  customer_name?: ModelStringInput | null,
  customer_image?: ModelStringInput | null,
  tripsID?: ModelIDInput | null,
  driverID?: ModelIDInput | null,
  userID?: ModelIDInput | null,
  and?: Array< ModelRatingFilterInput | null > | null,
  or?: Array< ModelRatingFilterInput | null > | null,
  not?: ModelRatingFilterInput | null,
  _deleted?: ModelBooleanInput | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type ModelTripsFilterInput = {
  id?: ModelIDInput | null,
  trip_status?: ModelTripStatusInput | null,
  pickup_lat?: ModelStringInput | null,
  pickup_lng?: ModelStringInput | null,
  dropoff_lat?: ModelStringInput | null,
  dropoff_lng?: ModelStringInput | null,
  pickup_location_description?: ModelStringInput | null,
  dropoff_location_description?: ModelStringInput | null,
  trip_cost?: ModelFloatInput | null,
  TotalQtyItems?: ModelIntInput | null,
  receipt?: ModelStringInput | null,
  duration?: ModelFloatInput | null,
  distance?: ModelFloatInput | null,
  store_name?: ModelStringInput | null,
  store_image?: ModelStringInput | null,
  user_name?: ModelStringInput | null,
  user_photo?: ModelStringInput | null,
  mobile_number?: ModelStringInput | null,
  driver_name?: ModelStringInput | null,
  driver_photo?: ModelStringInput | null,
  driver_rating?: ModelFloatInput | null,
  driver_mobile_number?: ModelStringInput | null,
  driver_car_model?: ModelStringInput | null,
  driver_car_make?: ModelStringInput | null,
  driver_car_color?: ModelStringInput | null,
  driver_car_plate_number?: ModelStringInput | null,
  sub_total?: ModelFloatInput | null,
  redeem_points?: ModelIntInput | null,
  service_fee?: ModelFloatInput | null,
  delivery_fee?: ModelFloatInput | null,
  delivery_discount?: ModelFloatInput | null,
  sum_total?: ModelFloatInput | null,
  userID?: ModelIDInput | null,
  storeID?: ModelIDInput | null,
  and?: Array< ModelTripsFilterInput | null > | null,
  or?: Array< ModelTripsFilterInput | null > | null,
  not?: ModelTripsFilterInput | null,
  _deleted?: ModelBooleanInput | null,
  tripsDriverId?: ModelIDInput | null,
};
// ...
export type RatingsByDriverIDQueryVariables = {
  driverID: string,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelRatingFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type RatingsByDriverIDQuery = {
  ratingsByDriverID?:  {
    __typename: "ModelRatingConnection",
    items:  Array< {
      __typename: "Rating",
      id: string,
      rating?: number | null,
      comment?: string | null,
      driver_image?: string | null,
      driver_name?: string | null,
      customer_name?: string | null,
      customer_image?: string | null,
      tripsID: string,
      driverID: string,
      userID: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};
// ...
export type ListTripsQueryVariables = {
  filter?: ModelTripsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListTripsQuery = {
  listTrips?:  {
    __typename: "ModelTripsConnection",
    items:  Array< {
      __typename: "Trips",
      id: string,
      trip_status?: TripStatus | null,
      pickup_lat?: string | null,
      pickup_lng?: string | null,
      dropoff_lat?: string | null,
      dropoff_lng?: string | null,
      pickup_location_description?: string | null,
      dropoff_location_description?: string | null,
      trip_cost?: number | null,
      TotalQtyItems?: number | null,
      receipt?: string | null,
      duration?: number | null,
      distance?: number | null,
      store_name?: string | null,
      store_image?: string | null,
      user_name?: string | null,
      user_photo?: string | null,
      mobile_number?: string | null,
      driver_name?: string | null,
      driver_photo?: string | null,
      driver_rating?: number | null,
      driver_mobile_number?: string | null,
      driver_car_model?: string | null,
      driver_car_make?: string | null,
      driver_car_color?: string | null,
      driver_car_plate_number?: string | null,
      sub_total?: number | null,
      redeem_points?: number | null,
      service_fee?: number | null,
      delivery_fee?: number | null,
      delivery_discount?: number | null,
      sum_total?: number | null,
      Driver?:  {
        __typename: "Driver",
        id: string,
        name: string,
        email: string,
        image?: string | null,
        phone_number?: string | null,
        rating?: number | null,
        inviteCode?: string | null,
        lng?: string | null,
        lat?: string | null,
        verificationStatus?: DriverRegistrationStatus | null,
        onlineStatus?: boolean | null,
        fcmToken?: string | null,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null,
      userID?: string | null,
      storeID: string,
      TripCarts?:  {
        __typename: "ModelTripCartConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Ratings?:  {
        __typename: "ModelRatingConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      tripsDriverId?: string | null,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};